Enlarge Product Search on Sales Order
Script to enlarge the product search on sales order and remove some colums. Standard UNleashed search window is quote narrow and has many columns so seeing information easily can be difficult especially of descriptions are very long. The standard screen can have the columns in the fixed window stretched but these changes won't be saved for next uses. The script widens the overall search box and removes so columns but could be changed for user requirements